initial exceptions.py
