+++ manticore/core/executor.py
@@ -4,7 +4,7 @@ import random
 import logging
 import signal
 
-from ..manticore import ManticoreError
+from ..exceptions import ExecutorError
 from ..utils.nointerrupt import WithKeyboardInterruptAs
 from ..utils.event import Eventful
 from .smtlib import Z3Solver, Expression, SolverException
@@ -15,10 +15,6 @@ from multiprocessing.managers import SyncManager
 from contextlib import contextmanager
 
 # This is the single global manager that will handle all shared memory among workers
-
-
-class ExecutorError(ManticoreError):
-    pass
 
 
 def mgr_init():
